+++ src/main/java/core/db/HRFTable.java
@@ -131,9 +131,6 @@ public final class HRFTable extends AbstractTable {
 	}
 
 	private HRF loadMaxHrf() {
-		/**
-		 * liefert die Maximal Vergebene Id eines HRF-Files
-		 */
 		String loadMaxHrfSql = createSelectStatement("order by HRF_ID desc LIMIT 1");
 		return loadHRF(loadMaxHrfSql);
 	}
@@ -157,7 +154,7 @@ public final class HRFTable extends AbstractTable {
 	}
 
 	public List<Integer> getHrfIdPerWeekList(int nWeeks) {
-		var sql = "select min(hrf_id) as id from " +
+		var sql = "select max(hrf_id) as id from " +
 				getTableName() +
 				" group by unix_timestamp(datum)/7/86400 order by id desc limit " +
 				nWeeks;