+++ src/main/resources/release_notes.md
@@ -1,6 +1,7 @@
 
 
 ## Highlights
+* Calculation of the tsi sub by approximating the tsi formula (#235)
 
 
 ## [Detailed Changelog](https://github.com/ho-dev/HattrickOrganizer/issues?q=milestone%3A9.0)
@@ -9,6 +10,7 @@
 * Increase match report column size (#2065)
 
 ### Squad
+* Calculation of the tsi sub by approximating the tsi formula (#235)
 
 ### Team Analyzer
 
@@ -35,7 +37,7 @@
 
 ## Translations
 
-Reports by Contributors - April 28, 2024 - May 06, 2024
+Reports by Contributors - April 28, 2024 - June 13, 2024
 
 * wsbrenk 0